Changed BoardType to be player indexed.

+++ src/games/tic_tac_toe/tic_tac_toe_state.cpp
@@ -4,16 +4,17 @@
 
 TicTacToeState::TicTacToeState() = default;
 
-void TicTacToeState::print_board() {
+void TicTacToeState::print_board()
+{
     // Print the board to screen.
     // LSB of the bit representation is top-left cell;
     // MSB of the bit representation is bottom-right.
     for (int row = 0; row < 3; row++) {
         for (int col = 0; col < 3; col++) {
             int bit = (row * 3) + col;
-            if ((board_[static_cast<int>(Player::One)] >> bit) & 1)
+            if ((board_[Player::One] >> bit) & 1)
                 std::cout << GREEN << "x " << RESET;
-            else if ((board_[static_cast<int>(Player::Two)] >> bit) & 1)
+            else if ((board_[Player::Two] >> bit) & 1)
                 std::cout << RED << "o " << RESET;
             else
                 std::cout << ". ";
@@ -25,36 +26,41 @@ void TicTacToeState::print_board() {
 
 void TicTacToeState::set_board(BoardType board) { this->board_ = board; }
 
-std::string TicTacToeState::state_to_string() {
+// void TicTacToeState::set_board(std::array<std::uint16_t, 2> board) { 
+//     this->board_[TicTacToeState::Player::One] = board[0]; 
+//     this->board_[TicTacToeState::Player::Two] = board[1]; 
+// }
+
+std::string TicTacToeState::state_to_string()
+{
     // Converts the state representation to readable string.
     // First nine characters represent the board.
     // Last character is the current player at the state.
     std::string state_str = "";
     for (int i = 0; i < 9; i++) {
-        if (board_[static_cast<int>(Player::One)] & 1)
+        if (board_[Player::One] & 1)
             state_str += '1';
-        else if (board_[static_cast<int>(Player::Two)] & 1)
+        else if (board_[Player::Two] & 1)
             state_str += '2';
         else
             state_str += '0';
-        board_[static_cast<int>(Player::One)] =
-            (board_[static_cast<int>(Player::One)] >> 1);
-        board_[static_cast<int>(Player::Two)] =
-            (board_[static_cast<int>(Player::Two)] >> 1);
+        board_[Player::One] = (board_[Player::One] >> 1);
+        board_[Player::Two] = (board_[Player::Two] >> 1);
     }
     state_str += static_cast<char>(player_);
     return state_str;
 }
 
-void TicTacToeState::string_to_state(std::string state_str) {
-    board_[static_cast<int>(Player::One)] = 0;
-    board_[static_cast<int>(Player::Two)] = 0;
+void TicTacToeState::string_to_state(std::string state_str)
+{
+    board_[Player::One] = 0;
+    board_[Player::Two] = 0;
     int count = 0;
     for (int i = 0; i < 9; i++) {
         if (state_str[i] == '1')
-            (board_[static_cast<int>(Player::One)] |= (1L << count));
+            (board_[Player::One] |= (1L << count));
         else if (state_str[i] == '2')
-            (board_[static_cast<int>(Player::Two)] |= (1L << count));
+            (board_[Player::Two] |= (1L << count));
         count++;
     }
     if (state_str.back() == '0')
